[feat] add ControllerScanner
