+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -2,7 +2,7 @@ package nextstep.mvc;
 
 import nextstep.mvc.exception.HandlerAdapterNotFoundException;
 import nextstep.mvc.exception.HandlerNotExistException;
-import nextstep.mvc.tobe.HandlerAdaptor;
+import nextstep.mvc.tobe.HandlerAdapter;
 import nextstep.mvc.tobe.ModelAndView;
 import nextstep.mvc.tobe.View;
 import nextstep.mvc.tobe.ViewResolver;
@@ -22,11 +22,11 @@ import java.util.Objects;
 public class DispatcherServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
     private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
-    private final List<HandlerAdaptor> handlerAdaptors;
+    private final List<HandlerAdapter> handlerAdapters;
     private final List<HandlerMapping> handlerMappings;
 
-    public DispatcherServlet(List<HandlerAdaptor> handlerAdaptors, List<HandlerMapping> handlerMappings) {
-        this.handlerAdaptors = handlerAdaptors;
+    public DispatcherServlet(List<HandlerAdapter> handlerAdapters, List<HandlerMapping> handlerMappings) {
+        this.handlerAdapters = handlerAdapters;
         this.handlerMappings = handlerMappings;
     }
 
@@ -41,9 +41,9 @@ public class DispatcherServlet extends HttpServlet {
         logger.debug("Method : {}, Request URI : {}", req.getMethod(), requestUri);
 
         Object handler = getHandler(req);
-        HandlerAdaptor handlerAdaptor = findHandlerAdapter(handler);
+        HandlerAdapter handlerAdapter = findHandlerAdapter(handler);
         try {
-            ModelAndView modelAndView = handlerAdaptor.handle(req, resp, handler);
+            ModelAndView modelAndView = handlerAdapter.handle(req, resp, handler);
             view(req, resp, modelAndView);
         } catch (Exception e) {
             logger.info("URI: {}", req.getRequestURI());
@@ -58,9 +58,9 @@ public class DispatcherServlet extends HttpServlet {
         view.render(modelAndView.getModel(), req, resp);
     }
 
-    private HandlerAdaptor findHandlerAdapter(Object handler) {
-        return handlerAdaptors.stream()
-            .filter(handlerAdaptor -> handlerAdaptor.supports(handler))
+    private HandlerAdapter findHandlerAdapter(Object handler) {
+        return handlerAdapters.stream()
+            .filter(handlerAdapter -> handlerAdapter.supports(handler))
             .findAny()
             .orElseThrow(HandlerAdapterNotFoundException::new);
     }
