+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/AnnotationHandlerMapping.java
@@ -29,44 +29,9 @@ public class AnnotationHandlerMapping implements HandlerMapping {
 
     @Override
     public void initialize() {
-        Reflections reflections = new Reflections(basePackage);
-        scanControllers(reflections);
+        ControllerScanner controllerScanner = new ControllerScanner(new Reflections(basePackage));
+        handlerExecutions = controllerScanner.scan();
         // TODO scan services
-    }
-
-    private void scanControllers(Reflections reflections) {
-        Set<Class<?>> controllers = reflections.getTypesAnnotatedWith(Controller.class);
-        try {
-            for (Class<?> controller : controllers) {
-                findHandlersInController(controller);
-            }
-        } catch (Exception e) {
-            throw new AnnotationScanFailException(e.getCause());
-        }
-    }
-
-    private void findHandlersInController(Class<?> controller) throws InstantiationException, IllegalAccessException, InvocationTargetException, NoSuchMethodException {
-        Object controllerInstance = controller.getDeclaredConstructor().newInstance();
-        logger.debug("c : {}", controller);
-        Method[] methods = controller.getDeclaredMethods();
-        Arrays.stream(methods)
-            .filter(method -> method.isAnnotationPresent(RequestMapping.class))
-            .forEach(method -> addHandlerExcution(controllerInstance, method));
-    }
-
-    private void addHandlerExcution(Object controllerInstance, Method method) {
-        RequestMapping annotation = method.getAnnotation(RequestMapping.class);
-        String url = annotation.value();
-        RequestMethod[] requestMethods = annotation.method();
-
-        addHandlerKey(controllerInstance, method, url, requestMethods);
-    }
-
-    private void addHandlerKey(Object controllerInstance, Method method, String url, RequestMethod[] requestMethods) {
-        for (RequestMethod requestMethod : requestMethods) {
-            HandlerKey key = new HandlerKey(url, requestMethod);
-            handlerExecutions.put(key, (req, res) -> method.invoke(controllerInstance, req, res));
-        }
     }
 
     @Override
